Buttons are DONE MOTHERFUCKAAAA
But seriously, that was gruelling.  The perfectionist in me says there's still more work to be done to tidy it up, but it's actually pretty good.  The only thing that's still kind of a little glitchy is if you scroll to the bottom of the screen too fast (which I'm planning to make the button do anyway) the Contact button doesn't align all the way across with the other buttons, you need to just scroll up a fraction to get it to stick.  Possibly there's another way to check if scroll is at bottom or manually stick it where it should be in that last frame of scroll.  But, for now, I'm moving on to the next section, which is trees in the wind...