Started work on the projects section
This is a copy of the menu I used in MonsterBasket, but rewriting as much as I can from memory to make sure I actually DO remember it all. Progress has been pretty quick so far, but I did need to look it up to see how I went about swapping classes.