Too many little things for one commit
Mostly just tidy work, but I've also swapped the contents of the about and contact sections.  And I'm not particularly happy with either.  The About section, which is now the one with the leaves is too plain.  I like the way the tree worked technically, but it was ugly, not sure what to do here.  Also don't know what to do with the Contact section, which is now the one with the sliders.  The bottom half is bare.